requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
Appium-Python-Client==2.8.1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
et-xmlfile==1.1.0
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
py==1.11.0
PySocks==1.7.1
pytest==7.2.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-xdist==3.1.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
python-dotenv==0.21.1
pandas~=1.5.3